// src/app/shared/models/services.model.ts
export class Services {
  group_id?: number = 0;
  app_key: string = '';
  title: string = '';
  descriptions: string = '';
  parent_id: string = '';
  parents_id: string = '';
  levels: number = 0;
  images: string = '';
  icons: string = '';
  quantity: number = 0;
  positions: string = '';
  orders: number = 0;
  created_by: string = '';
  created_at: Date = new Date();
  updated_by: string = '';
  updated_at: Date = null;
  deleted_by: string = '';
  deleted_at: Date = null;
  flag: number = 1;
}